+++ test/PolynomialsTests/QCurlGradBasesTests.jl
@@ -1,4 +1,4 @@
-module QCurlGradMonomialBasesTests
+module QCurlGradBasesTests
 
 using Test
 using Gridap.TensorValues
@@ -14,7 +14,7 @@ D = 2
 T = Float64
 V = VectorValue{D,T}
 G = gradient_type(V,xi)
-b = QCurlGradMonomialBasis(Val(D),T,order)
+b = QCurlGradBasis(Monomial, Val(D),T,order)
 
 @test length(b) == 4
 @test get_order(b) == 1
@@ -28,7 +28,7 @@ D = 3
 T = Float64
 V = VectorValue{D,T}
 G = gradient_type(V,xi)
-b = QCurlGradMonomialBasis(Val(D),T,order)
+b = QCurlGradBasis(Monomial, Val(D),T,order)
 
 v = V[
   (1.0, 0.0, 0.0), (2.0, 0.0, 0.0),
@@ -57,7 +57,7 @@ D = 2
 T = Float64
 V = VectorValue{D,T}
 G = gradient_type(V,xi)
-b = QCurlGradMonomialBasis(Val(D),T,order)
+b = QCurlGradBasis(Monomial, Val(D),T,order)
 
 v = V[
   ( 1., 0. ), ( 2., 0. ), ( 4., 0. ), ( 3., 0. ), ( 6., 0. ), (12., 0. ),