+++ test/PolynomialsTests/QGradBasesTests.jl
@@ -15,7 +15,7 @@ D = 2
 T = Float64
 V = VectorValue{D,T}
 G = gradient_type(V,xi)
-b = QGradMonomialBasis(Val(D),T,order)
+b = QGradBasis(Monomial, Val(D),T,order)
 
 @test length(b) == 4
 @test get_order(b) == 1
@@ -29,7 +29,7 @@ D = 3
 T = Float64
 V = VectorValue{D,T}
 G = gradient_type(V,xi)
-b = QGradMonomialBasis(Val(D),T,order)
+b = QGradBasis(Monomial, Val(D),T,order)
 
 v = V[
   (1.0, 0.0, 0.0), (3.0, 0.0, 0.0), (5.0, 0.0, 0.0), (15.0, 0.0, 0.0),
@@ -68,7 +68,7 @@ D = 2
 T = Float64
 V = VectorValue{D,T}
 G = gradient_type(V,xi)
-b = QGradMonomialBasis(Val(D),T,order)
+b = QGradBasis(Monomial, Val(D),T,order)
 
 v = V[
   (1.0, 0.0), (0.0, 1.0), (2.0, 0.0), (0.0, 3.0),
@@ -94,43 +94,6 @@ g = G[
   (0., 0., 3., 2.),
   (0., 0.,12., 4.)]
 
-bx = repeat(permutedims(v),np)
-∇bx = repeat(permutedims(g),np)
-test_field_array(b,x,bx,grad=∇bx)
-test_field_array(b,x[1],bx[1,:],grad=∇bx[1,:])
-
-xi = Point(2.,3.,5.)
-np = 3
-x = fill(xi,np)
-
-order = 0
-D = 3
-b = Polynomials.NedelecPrebasisOnSimplex{D}(order)
-
-V = VectorValue{D, Float64}
-v = V[(1,0,0),(0,1,0),(0,0,1),(-3,2,0),(-5,0,2),(0,-5,3)]
-
-G = gradient_type(V,xi)
-g = G[
-  (0,0,0, 0,0,0, 0,0,0), (0,0,0, 0,0,0, 0,0,0), (0,0,0, 0,0,0, 0,0,0),
-  (0,-1,0, 1,0,0, 0,0,0),(0,0,-1, 0,0,0, 1,0,0),(0,0,0, 0,0,-1, 0,1,0)]
-
-bx = repeat(permutedims(v),np)
-∇bx = repeat(permutedims(g),np)
-test_field_array(b,x,bx,grad=∇bx)
-test_field_array(b,x[1],bx[1,:],grad=∇bx[1,:])
-
-xi = Point(2.,3.)
-np = 4
-x = fill(xi,np)
-
-order = 0
-D = 2
-b = Polynomials.NedelecPrebasisOnSimplex{D}(order)
-V = VectorValue{D, Float64}
-v = V[(1,0),(0,1),(-3,2)]
-G = gradient_type(V,xi)
-g = G[(0,0, 0,0), (0,0, 0,0), (0,-1, 1,0)]
 bx = repeat(permutedims(v),np)
 ∇bx = repeat(permutedims(g),np)
 test_field_array(b,x,bx,grad=∇bx)